Simplify simple tensor types, isolate OrtValueExtensions
